+++ midas-portal/src/app/components/edi-list/edi-list-modal.component.html
@@ -1,7 +1,6 @@
-<h1 mat-dialog-title>EDI List</h1>
 <div mat-dialog-content>
-  <p>Table here</p>
-  
+  <app-edi-list></app-edi-list>
+
 </div>
 <div mat-dialog-actions>
   <button mat-button (click)="onNoClick()">Close</button>
